+++ src/Device/DeviceSelector/DeviceSelector.tsx
@@ -4,12 +4,17 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import React, { FC, useCallback, useEffect, useState } from 'react';
+import React, { useCallback, useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { DeviceTraits } from '@nordicsemiconductor/nrf-device-lib-js';
 
 import { Device } from '../../state';
 import useHotKey from '../../utils/useHotKey';
+import {
+    clearWaitForDeviceTimeout,
+    getWaitingToAutoReselect,
+    setAutoSelectDevice,
+} from '../deviceAutoSelectSlice';
 import { startWatchingDevices, stopWatchingDevices } from '../deviceLister';
 import { DeviceSetup as DeviceSetupShared, setupDevice } from '../deviceSetup';
 import DeviceSetup from '../DeviceSetup/DeviceSetup';
@@ -22,8 +27,6 @@ import {
 import DeviceList from './DeviceList/DeviceList';
 import SelectDevice from './SelectDevice';
 import SelectedDevice from './SelectedDevice';
-import useAutoReconnectDevice from './useAutoReconnectDevice';
-import useAutoselectDevice from './useAutoselectDevice';
 
 interface OutdatedDeviceTraits {
     serialPort?: boolean;
@@ -34,7 +37,7 @@ export interface Props {
     deviceListing: DeviceTraits & OutdatedDeviceTraits;
     deviceSetup?: DeviceSetupShared;
     releaseCurrentDevice?: () => void;
-    onDeviceSelected?: (device: Device) => void;
+    onDeviceSelected?: (device: Device, autoReselected: boolean) => void;
     onDeviceDeselected?: () => void;
     onDeviceConnected?: (device: Device) => void;
     onDeviceDisconnected?: (device: Device) => void;
@@ -43,7 +46,7 @@ export interface Props {
 }
 
 const noop = () => {};
-const DeviceSelector: FC<Props> = ({
+export default ({
     deviceListing,
     deviceSetup,
     releaseCurrentDevice = noop,
@@ -53,17 +56,60 @@ const DeviceSelector: FC<Props> = ({
     onDeviceDisconnected = noop,
     onDeviceIsReady = noop,
     deviceFilter,
-}) => {
+}: Props) => {
     const dispatch = useDispatch();
     const [deviceListVisible, setDeviceListVisible] = useState(false);
 
     const deviceIsSelected = useSelector(deviceIsSelectedSelector);
     const selectedSN = useSelector(selectedSerialNumber);
+    const waitingToAutoReconnect = useSelector(getWaitingToAutoReselect);
+    const showSelectedDevice = deviceIsSelected || waitingToAutoReconnect;
 
     const doDeselectDevice = useCallback(() => {
         onDeviceDeselected();
         dispatch(deselectDevice());
+        dispatch(setAutoSelectDevice(undefined));
     }, [dispatch, onDeviceDeselected]);
+
+    const doSelectDevice = useCallback(
+        (device: Device, autoReselected: boolean) => {
+            if (device.serialNumber === selectedSN) {
+                setDeviceListVisible(false);
+                return;
+            }
+
+            if (deviceIsSelected) {
+                doDeselectDevice();
+            }
+
+            dispatch(clearWaitForDeviceTimeout());
+            setDeviceListVisible(false);
+            dispatch(selectDevice(device));
+            dispatch(setAutoSelectDevice(device));
+            onDeviceSelected(device, autoReselected);
+            if (deviceSetup) {
+                dispatch(
+                    setupDevice(
+                        device,
+                        deviceSetup,
+                        releaseCurrentDevice,
+                        onDeviceIsReady,
+                        doDeselectDevice
+                    )
+                );
+            }
+        },
+        [
+            deviceIsSelected,
+            deviceSetup,
+            dispatch,
+            doDeselectDevice,
+            onDeviceIsReady,
+            onDeviceSelected,
+            releaseCurrentDevice,
+            selectedSN,
+        ]
+    );
 
     const doStartWatchingDevices = useCallback(() => {
         const patchedDeviceListing = {
@@ -75,7 +121,8 @@ const DeviceSelector: FC<Props> = ({
                 patchedDeviceListing,
                 onDeviceConnected,
                 onDeviceDisconnected,
-                onDeviceDeselected
+                onDeviceDeselected,
+                doSelectDevice
             )
         );
     }, [
@@ -84,33 +131,8 @@ const DeviceSelector: FC<Props> = ({
         onDeviceConnected,
         onDeviceDisconnected,
         onDeviceDeselected,
+        doSelectDevice,
     ]);
-
-    const doSelectDevice = (device: Device) => {
-        if (device.serialNumber === selectedSN) {
-            setDeviceListVisible(false);
-            return;
-        }
-
-        if (deviceIsSelected) {
-            doDeselectDevice();
-        }
-        setDeviceListVisible(false);
-        dispatch(selectDevice(device));
-        onDeviceSelected(device);
-        if (deviceSetup) {
-            dispatch(
-                setupDevice(
-                    device,
-                    deviceSetup,
-                    releaseCurrentDevice,
-                    onDeviceIsReady,
-                    doStartWatchingDevices,
-                    doDeselectDevice
-                )
-            );
-        }
-    };
 
     const toggleDeviceListVisible = () =>
         setDeviceListVisible(!deviceListVisible);
@@ -119,9 +141,6 @@ const DeviceSelector: FC<Props> = ({
         doStartWatchingDevices();
         return stopWatchingDevices;
     }, [doStartWatchingDevices]);
-
-    useAutoselectDevice(doSelectDevice);
-    useAutoReconnectDevice(doSelectDevice);
 
     useHotKey({
         hotKey: 'alt+s',
@@ -132,7 +151,7 @@ const DeviceSelector: FC<Props> = ({
 
     return (
         <div className="core19-device-selector">
-            {deviceIsSelected ? (
+            {showSelectedDevice ? (
                 <SelectedDevice
                     doDeselectDevice={doDeselectDevice}
                     toggleDeviceListVisible={toggleDeviceListVisible}
@@ -152,5 +171,3 @@ const DeviceSelector: FC<Props> = ({
         </div>
     );
 };
-
-export default DeviceSelector;